+++ examples/README.md
@@ -117,8 +117,7 @@ relevant code.
 - [SIR Epidemiology model](./toy/model-sir.ipynb)
 - [Stochastic Degradation model](./toy/model-stochastic-degradation.ipynb)
 - [Stochastic Logistic model](./toy/model-stochastic-logistic-growth.ipynb)
-- [Michaelis Menten model](./toy/model-michaelis-menten.ipynb)
-- [Degradation model](./toy/model-degradation.ipynb)
+- [Michaelis Menten model](./toy/model-stochastic-michaelis-menten.ipynb)
 
 ### Distributions
 - [Annulus](./toy/distribution-annulus.ipynb)
